% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignMarineAreas.R
\name{assignMarineAreas}
\alias{assignMarineAreas}
\title{assignMarineAreas}
\usage{
assignMarineAreas(data, LocGroup1, LocGroup2, LocGroup3)
}
\arguments{
\item{data}{Filtered permit data frame, with the added \code{Location} column already
created.}

\item{LocGroup1}{Character vector describing Location names to assign to HUC 8 number
17110019 - "Puget Sound".}

\item{LocGroup2}{Character vector describing Location names to assign to HUC 8 number
17110018 - "Hood Canal".}

\item{LocGroup3}{Character vector describing Location names to assign to HUC 8 number
17110021 - "Crescent-Hoko".Note that it is possible that a different HUC could have been
chosen (17110020 - "Dungeness-Elwha") as either of these HUCs can describe Strait of Juan
de Fuca research.}
}
\value{
A very similar data frame, know with more HUC's assigned to marine areas
}
\description{
Assigns HUC 8 codes to marine areas based on Location column
}
