% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignWaterType.R
\name{assignWaterType}
\alias{assignWaterType}
\title{assignWaterType}
\usage{
assignWaterType(x, strings1 = SW_strings, strings2 = FW_strings)
}
\arguments{
\item{x}{a character column of Location names.}

\item{strings1}{A vector of common strings found in the name of saltwater areas. Strings
can be user created, or default sting names can be used that are stored within the package.
The list of default saltwater stings can be accessed by calling \code{SW_strings}.}

\item{strings2}{A vector of common strings found in the name of freshwater areas. Strings
can be user created, or default string names can be used that are stored within the package.
The list of default freshwater strings can be accessed by calling \code{FW_strings}.}
}
\value{
A character vector denoting freshwater, saltwater, or both.
}
\description{
Creates a vector describing whether the research permits are occuring
in freshwater or saltwater based on strings within the Location column.
}
