% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ESUBoundary.R
\name{create_ESUBoundary}
\alias{create_ESUBoundary}
\title{create_ESUBoundary}
\usage{
create_ESUBoundary(data, spatialData)
}
\arguments{
\item{data}{Data frame describing the different HUC 8 where ESUs can be found. One row per unique
DPS/huc8 combination. Therefore species that are found in multiple DPSs wil be lsted in multiple
rows, one for each DPS where they can be found.}

\item{spatialData}{Polygon spatial data for HUC 8's in Washington, Idaho, Oregon and California.
Recommended to use the data that comes with this package from the Watershed Boundary Dataset (Made
and maintained by the USGS).}
}
\value{
A large multipolygon data frame, where each ESU/DPS has a collection of spatial polygons that
show the full extent of where the ESU may be found.
}
\description{
Filters to only include species relevant to the app.
}
