% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLocations.R
\name{createLocations}
\alias{createLocations}
\title{createLocations}
\usage{
createLocations(data)
}
\arguments{
\item{data}{Filtered permit data frame, usually the result of \code{permitFilter} function}
}
\value{
A data frame of research permits, but now with the added Location column.
}
\description{
Creates a Location field that describes the location of the research, based on
hierarchical sort of the 4 location-based columns in the raw data. The default hierarchy is
WaterbodyName > BasinName > StreamName > LocationDescription, where the highest-level data that
is not NA was used. For example, if a permit had an NA value for WaterbodyName, and a non-NA value
for BasinName, the BasinName value would be used to describe that permits location in the new “Location” column.
}
