% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMapDF.R
\name{createMapDF}
\alias{createMapDF}
\title{createMapDF}
\usage{
createMapDF(data, spatialData, createTakeVar = T)
}
\arguments{
\item{data}{Filtered permit data frame with TotalMorts column already created.}

\item{spatialData}{Polygon spatial data for HUC 8's in Washington, Idaho, Oregon and California.
Recommended to use the data that comes with this package from the Watershed Boundary Dataset (Made
and maintained by the USGS).}

\item{createTakeVar}{Logical indicating whether total take (sum of lethal and non-lethal take) data
should be summarized instead of only lethal take data. Defaults to TRUE indicating that total take
data should be summarized}
}
\value{
A spatial data frame with the total amount of take or mortalities for each unique ESU, huc8,
LifeStage, and Production combination.
}
\description{
Creates data frames for input into Leaflet map.
}
