% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esuBoundaryFilter.R
\name{esuBoundaryFilter}
\alias{esuBoundaryFilter}
\title{esuBoundaryFilter}
\usage{
esuBoundaryFilter(data)
}
\arguments{
\item{data}{Data frame describing the different HUC 8 where ESUs can be found}
}
\value{
A data frame that contains only the relevant species.
}
\description{
Filters to only include species relevant to the app.
}
