% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeLeafletApp.R
\name{MakeLeafletApp}
\alias{MakeLeafletApp}
\title{MakeLeafletApp}
\usage{
MakeLeafletApp(DF, spatialData, ESUhucData)
}
\arguments{
\item{DF}{Data frame describing the different HUC 8 where ESUs can be found. One row per unique
DPS/huc8 combination. Therefore species that are found in multiple DPSs wil be lsted in multiple
rows, one for each DPS where they can be found.}

\item{spatialData}{Polygon spatial data for HUC 8's in Washington, Idaho, Oregon and California.
Recommended to use the data that comes with this package from the Watershed Boundary Dataset (Made
and maintained by the USGS).}

\item{ESUhucData}{Data frame describing the different HUC 8 where ESUs can be found. One row per unique
DPS/huc8 combination. Therefore species that are found in multiple DPSs wil be lsted in multiple
rows, one for each DPS where they can be found.}
}
\value{
A Shiny application that displays take data in a Leaflet map, as well as a data table showing
what makes up the take values.
}
\description{
Creates the Leaflet app displaying take data for ESA-listed species.
}
